requirements.txt
langchain>=0.1.0
langchain-community
langchain-huggingface
transformers>=4.38.0
sentence-transformers>=2.2.2
huggingface-hub>=0.18.0
qdrant-client>=1.6.0
diffusers>=0.25.0
torch>=2.0.0
pandas>=1.5.3
fastapi>=0.100
uvicorn>=0.23
python-dotenv>=1.0.0
email-validator>=2.1.0
pytest
strawberry-graphql[fastapi]
duckduckgo-search
sacremoses
google-search-results